MAIHDA: Multilevel Analysis of Individual Heterogeneity and Discriminatory Accuracy

Provides a comprehensive toolkit for conducting Multilevel Analysis of Individual Heterogeneity and Discriminatory Accuracy (MAIHDA). Methods are described in Merlo (2018) doi:10.1016/j.socscimed.2017.12.018 and Evans et al. (2018) doi:10.1016/j.socscimed.2017.11.011 . Automatically generates intersectional strata, fits analytical models, extracts statistics, and produces visualizations.

See also

Author

Maintainer: Hamid Bulut me@hamidbulut.com